Extend integer literals to ulong to allow for -2^63 - 2^(63-1) literal range

+++ Toffee/LexicalAnalysis/Lexer_Numbers.cs
@@ -6,7 +6,7 @@ public sealed partial class Lexer
 {
     private Token? MatchNumber()
     {
-        // TODO: -9223372036854775808 would not be matched correctly
+        // TODO: handle <-9223372036854775808 and >9223372036854775808 "literals" in parser
         if (!IsDigitGivenRadix(10, _scanner.CurrentCharacter))
             return null;
 
@@ -22,12 +22,12 @@ public sealed partial class Lexer
     private Token ContinueMatchingDecimalNumber(char initialDigit)
     {
         static bool IsDigit(char? c) => IsDigitGivenRadix(10, c);
-        void AppendDigitConsideringOverflow(ref long buffer, char digit, ref bool overflowOccurred, Position? errorPosition = null) =>
+        void AppendDigitConsideringOverflow(ref ulong buffer, char digit, ref bool overflowOccurred, Position? errorPosition = null) =>
             AppendDigitConsideringOverflowGivenRadix(10, ref buffer, digit, ref overflowOccurred, errorPosition);
 
         var overflowOccurred = false;
 
-        var integralPart = (long)CharToDigit(initialDigit);
+        var integralPart = (ulong)CharToDigit(initialDigit);
         while (IsDigit(_scanner.CurrentCharacter))
         {
             AppendDigitConsideringOverflow(ref integralPart, _scanner.CurrentCharacter!.Value, ref overflowOccurred);
@@ -37,7 +37,7 @@ public sealed partial class Lexer
             return new Token(TokenType.LiteralInteger, integralPart);
 
         _scanner.Advance();
-        var fractionalPart = 0L;
+        var fractionalPart = 0ul;
         var fractionalPartLength = 0;
         while (IsDigit(_scanner.CurrentCharacter))
         {
@@ -53,7 +53,7 @@ public sealed partial class Lexer
         }
 
         _scanner.Advance();
-        var exponentialPart = 0L;
+        var exponentialPart = 0ul;
         var exponentSign = _scanner.CurrentCharacter is '-' ? -1 : 1;
         if (_scanner.CurrentCharacter is '-' or '+')
             _scanner.Advance();
@@ -64,8 +64,8 @@ public sealed partial class Lexer
             AppendDigitConsideringOverflow(ref exponentialPart, _scanner.CurrentCharacter!.Value, ref overflowOccurred);
             _scanner.Advance();
         }
-        var exponentiatedNumber = integralPart * Math.Pow(10, exponentSign * exponentialPart)
-            + fractionalPart * Math.Pow(10, exponentSign * exponentialPart - fractionalPartLength);
+        var exponentiatedNumber = integralPart * Math.Pow(10, exponentSign * (double)exponentialPart)
+            + fractionalPart * Math.Pow(10, exponentSign * (double)exponentialPart - fractionalPartLength);
         return new Token(TokenType.LiteralFloat, exponentiatedNumber);
     }
 
@@ -81,18 +81,18 @@ public sealed partial class Lexer
         };
 
         bool IsDigit(char? c) => IsDigitGivenRadix(radix, c);
-        void AppendDigitConsideringOverflow(ref long buffer, char digit, ref bool overflowOccurred, Position? errorPosition = null) =>
+        void AppendDigitConsideringOverflow(ref ulong buffer, char digit, ref bool overflowOccurred, Position? errorPosition = null) =>
             AppendDigitConsideringOverflowGivenRadix(radix, ref buffer, digit, ref overflowOccurred, errorPosition);
 
         if (!IsDigit(_scanner.CurrentCharacter))
         {
             EmitError(new MissingNonDecimalDigits(_scanner.CurrentPosition, prefix));
-            return new Token(TokenType.LiteralInteger, 0L);
+            return new Token(TokenType.LiteralInteger, 0ul);
         }
 
         var overflowOccurred = false;
 
-        var integralPart = (long)CharToDigit(_scanner.CurrentCharacter!.Value);
+        var integralPart = (ulong)CharToDigit(_scanner.CurrentCharacter!.Value);
         _scanner.Advance();
         while (IsDigit(_scanner.CurrentCharacter))
         {
